requirements.txt
Django
boto
django-debug-toolbar
greeking
django-bakery
geopy